fix(cases): fix animation while installing special cases

+++ board/recalbox/fsoverlay/recalbox/scripts/recalbox-create-share.sh
@@ -1,5 +1,12 @@
 #!/bin/bash
 
+# Don't play animation when a special case is installing!
+if [[ -f /tmp/.install.png ]] ; then
+	exit 0
+fi
+
+touch /tmp/newshare
+
 dd if=/dev/zero of=/dev/fb0 > /dev/null 2>&1
 FILES=/recalbox/system/resources/creating-share*.jpg
 while true ; do
